Ajustement de web.php

+++ routes/web.php
@@ -1,11 +1,13 @@
 <?php
 
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\RepairController;
 use App\Http\Controllers\VehicleController;
 use App\Http\Controllers\UserController;
 //use App\Http\Controllers\Auth;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Foundation\Auth\EmailVerificationRequest;
 
 /*
 |--------------------------------------------------------------------------
@@ -21,52 +23,90 @@ use Illuminate\Support\Facades\Auth;
 
 
 Route::controller(RepairController::class)->group(function () {
-    Route::resource('repairs', RepairController::class);
+    //Route::resource('repairs', RepairController::class);
 
-    Route::get('/', 'index');
-    Route::get('/repair/create', 'create');
-    Route::get('/repair/{id}', 'show');
-    Route::get('/repair/{id}/edit', 'edit');
+    Route::get('/', 'index')->middleware(['auth','verified']);
+    Route::get('/repair/create', 'create')->middleware(['auth','verified']);
+    Route::get('/repair/{id}', 'show')->middleware(['auth','verified']);
+    Route::get('/repair/{id}/edit', 'edit')->middleware(['auth','verified']);
 
-    Route::post('/repair', 'store');
-    Route::patch('/repair/{id}', 'update');
-    Route::delete('/repair/{id}', 'destroy');
+    Route::post('/repair', 'store')->middleware(['auth','verified']);
+    Route::patch('/repair/{id}', 'update')->middleware(['auth','verified']);
+    Route::delete('/repair/{id}', 'destroy')->middleware(['auth','verified']);
 });
 
 Route::controller(VehicleController::class)->group(function () {
-    Route::resource('vehicles', VehicleController::class);
+    //Route::resource('vehicles', VehicleController::class);
 
-    Route::get('/vehicle', 'index');
-    Route::get('/vehicle/create', 'create');
-    Route::get('/vehicle/{id}', 'show');
-    Route::get('/vehicle/{id}/edit', 'edit');
+    Route::get('/vehicle', 'index')->middleware(['auth','verified']);
+    Route::get('/vehicle/create', 'create')->middleware(['auth','verified']);
+    Route::get('/vehicle/{id}', 'show')->middleware(['auth','verified']);
+    Route::get('/vehicle/{id}/edit', 'edit')->middleware(['auth','verified']);
 
-    Route::post('/vehicle', 'store');
-    Route::patch('/vehicle/{id}', 'update');
-    Route::delete('/vehicle/{id}', 'destroy');
+    Route::post('/vehicle', 'store')->middleware(['auth','verified']);
+    Route::patch('/vehicle/{id}', 'update')->middleware(['auth','verified']);
+    Route::delete('/vehicle/{id}', 'destroy')->middleware(['auth','verified']);
 });
 
 Route::controller(UserController::class)->group(function () {
-    Route::resource('users', UserController::class);
+    //Route::resource('users', UserController::class);
 
-    Route::get('/user', 'index');
-    Route::get('/user/create', 'create');
-    Route::get('/user/{id}', 'show');
+    Route::get('/user', 'index')->middleware(['auth','verified']);
+    Route::get('/user/create', 'create')->middleware(['auth','verified']);
+    Route::get('/user/{id}', 'show')->middleware(['auth','verified']);
     //Route::get('/user/{id}/edit', 'edit');
 
-    Route::post('/user', 'store');
+    Route::post('/user', 'store')->middleware(['auth','verified']);
     //Route::patch('/user/{id}', 'update');
     //Route::delete('/user/{id}', 'destroy');
 });
 
-//Dashboard
-Route::group(['middleware' => ['auth']], function() {
-    Route::group(['middleware' => ['verified']], function() {
-        Route::get('/dashboard', 'DashboardController@index')->name('dashboard.index');
-    });
-});
+
+Route::get('/dashboard', function () {
+    return view('dashboard');
+})->middleware(['auth','verified'])->name('dashboard');
+
+//require __DIR__.'/auth.php';
+
+Route::get('/email/verify', function () {
+    return view('auth.verify-email');
+})->middleware('auth')->name('verification.notice');
+
+Route::get('/email/verify/{id}/{hash}', function (EmailVerificationRequest $request) {
+    $request->fulfill();
+ 
+    return redirect('/home');
+})->middleware(['auth', 'signed'])->name('verification.verify');
+
+Route::post('/email/verification-notification', function (Request $request) {
+    $request->user()->sendEmailVerificationNotification();
+ 
+    return back()->with('message', 'Verification link sent!');
+})->middleware(['auth', 'throttle:6,1'])->name('verification.resend');
 
 Auth::routes();
+
+
+
+
+// Route::get('/email/verify', function () {
+//     return view('auth.verify-email');
+// })->middleware('auth')->name('verification.notice');
+
+// Route::get('/email/verify/{id}/{hash}', function (EmailVerificationRequest $request) {
+//     $request->fulfill();
+ 
+//     return redirect('/home');
+// })->middleware(['auth', 'signed'])->name('verification.verify');
+
+// Route::post('/email/verification-notification', function (Request $request) {
+//     $request->user()->sendEmailVerificationNotification();
+ 
+//     return back()->with('message', 'Verification link sent!');
+// })->middleware(['auth', 'throttle:6,1'])->name('verification.send');
+
+
+
 
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
